fix: rename `http` to `operate_http`
