+++ operate/operate_http/__init__.py
@@ -28,7 +28,7 @@ from starlette.requests import Request
 from starlette.responses import JSONResponse
 from starlette.types import Receive, Scope, Send
 
-from operate.http.exceptions import NotAllowed, ResourceException
+from operate.operate_http.exceptions import NotAllowed, ResourceException
 
 
 # pylint: disable=no-self-use
